Apply suggestions from Adrien's review
